backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.3.0
pydantic==2.5.3
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
email-validator>=2.0.0